+++ apps/scouts/src/components/boost/boostCMS/boostCMSForms/boostCMSIssueTo/BoostShareableQRCode.tsx
@@ -17,7 +17,7 @@ const BoostShareableQRCode: React.FC<{
     boostClaimLink: string;
 }> = ({ handleCloseModal, state, boostClaimLink }) => {
     const defaultTitle =
-        CATEGORY_TO_SUBCATEGORY_LIST?.[state?.basicInfo?.type].find(
+        CATEGORY_TO_SUBCATEGORY_LIST?.[state?.basicInfo?.type]?.find(
             options => options?.type === state?.basicInfo?.achievementType
         )?.title ?? '';